chore: consolidate providers #43
- consolidated providers
- consolidated tailwind indicator and session-info
- updated layout

Closes #43
